fix compile time errors but still have runtime errors
