+++ CrusherChess/src/moves.h
@@ -1,0 +1,14 @@
+#pragma once
+//move list
+typedef struct Moves
+{
+	int moves[255];
+	//could possibily use this?
+	//std::size(data);
+
+   // int legal[255];
+
+	//move count
+	int count;
+
+} Moves;
