+++ CrusherChess/src/typedefs.h
@@ -1,0 +1,2 @@
+#pragma once
+typedef unsigned long long U64;
